requirements.txt
pymupdf
transformers
torch
fpdf2
PyYAML
langdetect
sentencepiece
sacremoses
fitz